+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 
 ## 0.0.7
 
-* Update Readme
+* Add Col Alignment
+* Add Floating NavBar
 
 ## 0.0.6
 
@@ -21,11 +22,11 @@
 
 ## 0.0.5
 
-* update Readme
+* Update Readme
 
 ## 0.0.4
 
-* update Readme
+* Update Readme
 
 
 ## 0.0.3